% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg.R
\name{asg.est}
\alias{asg.est}
\title{Parametros mi, sigma e alpha da funcao densidade de probabilidade da distribuicao Alpha Skew Gaussian (ASG)}
\usage{
asg.est(
  dados,
  var,
  condicional,
  classe = NULL,
  dg = FALSE,
  de_densidade = NULL
)
}
\arguments{
\item{dados}{Dataframe onde estão a variável observada e a variável conhecida (condicional)}

\item{var}{String formato "nome da coluna", parâmetro indica coluna onde está a variável observada}

\item{condicional}{String no formato "nome da coluna", parâmetro indica coluna onde está a variavel conhecida}

\item{classe}{String no formato "valor", valor condicional conhecido}

\item{dg}{Lógico, default = FALSE, se TRUE função não recebe condicional.}

\item{de_densidade}{Lista, parâmetros passados quando asg.est é chama por dens.descritiva. Não passe esses parâmetros}
}
\value{
Retorna valores de mi, sigma e alpha para a função densidade de probabilidade da distribuição Alpha Skew Gaussian (ASG)
}
\description{
Estimação de parâmetros mi, sigma e alpha para a função densidade de probabilidade da distribuição Alpha Skew Gaussian (ASG)
}
\examples{

#Exemplo 1

asg.est(iris, "Petal.Length", "Species", "setosa")

#Exemplo 2

asg.est(iris, "Sepal.Length", dg = TRUE)

}
