% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.R
\name{disp.descritiva}
\alias{disp.descritiva}
\title{Diagramas de disperso a partir de uma informao conhecida}
\usage{
disp.descritiva(dados, condicional, vars)
}
\arguments{
\item{dados}{Dataframe com ao menos trs colunas, uma com a informao a priori e no mnimo outras duas com valores para coordenadas x e y}

\item{condicional}{Coluna em que est a informao conhecida (priori)}

\item{vars}{Lista com os nomes das duas colunas ( x e y, nessa ordem) que formaro as coordenadas do plano cartesiano}
}
\description{
Essa funo l um conjunto de dados e plota diagramas de disperso para observao do comportamento de uma determinada relo entre variveis de acordo com informaes a priori
}
\examples{

#Exemplo 1:
#função recebe: dataset iris,
#               info a priori "Species"
#               colunas para coordenadas x e y "Petal.Length" e "Petal.Width"

disp.descritiva(iris, "Species", c("Sepal.Length", "Sepal.Width"))

#Exemplo 2:
#função recebe: dataset mtcars,
#               info a priori "cyl"
#               colunas para coordenadas x e y "mpg" e "qsec"

disp.descritiva(mtcars, "cyl", c("mpg", "qsec"))

}
