% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq.descritiva}
\alias{freq.descritiva}
\title{Tabela de frequência}
\usage{
freq.descritiva(dados, condicional, vars)
}
\arguments{
\item{dados}{Dataframe com ao menos duas colunas, uma com a informação a priori e outra(s) com valores para plotar histograma e boxplot}

\item{condicional}{Coluna em que está a informação conhecida}

\item{vars}{Lista com nomes das colunas em que estão os valores com os quais serão plotados o gráfico de barras e construida a tabela de frequência}
}
\description{
Lê conjunto de dados e retorna uma tabela de frequência com observações, frequência absoluta e proporcional
}
\examples{

mtcars_ <- mtcars
mtcars_$am[mtcars_$am == 0] <- "automatico"
mtcars_$am[mtcars_$am == 1] <- "manual"
freq.descritiva(mtcars_, "am", c("cyl", "gear"))

}
