% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens.R
\name{dens.descritiva}
\alias{dens.descritiva}
\title{Sobreposição de histograma e curva da função de densidade}
\usage{
dens.descritiva(dados, condicional, vars)
}
\arguments{
\item{dados}{Dataframe com ao menos duas colunas, uma com a informação a priori e outra(s) com valores para plotar histograma e boxplot}

\item{condicional}{Coluna em que está a informação conhecida}

\item{vars}{Lista com nomes das colunas em que estão os valores com os quais serão plotados o histograma e curva de densidade}
}
\value{
Plot com histograma e curva de densidade
}
\description{
Lê um conjunto de dados e plota a sobreposição do seu histograma e curva de densidade para observar o comportamento da distribuição dos dados
}
\examples{

dens.descritiva(iris, "Species", c("Petal.Length", "Sepal.Length"))

}
