% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.R
\name{disp.descritiva}
\alias{disp.descritiva}
\title{Diagramas de dispersão a partir de uma informação conhecida}
\usage{
disp.descritiva(dados, condicional, vars)
}
\arguments{
\item{dados}{Dataframe com ao menos três colunas, uma com a informação a priori e no mínimo outras duas com valores para coordenadas x e y}

\item{condicional}{Coluna em que está a informação conhecida (priori)}

\item{vars}{Lista com os nomes das duas colunas ( x e y, nessa ordem) que formarão as coordenadas do plano cartesiano}
}
\value{
A função retorna uma lista com diagramas de dispersão combinados e individuais do conjunto de dados observado bem como as tabelas definidas por cada priori
}
\description{
Essa função lê um conjunto de dados e plota diagramas de dispersão para observação do comportamento de uma determinada relação entre variáveis de acordo com informações conhecidas a priori
}
\examples{

#Exemplo 1:
#função recebe: dataset iris,
#               info a priori "Species"
#               colunas para coordenadas x e y "Petal.Length" e "Petal.Width"

disp.descritiva(iris, "Species", c("Sepal.Length", "Sepal.Width"))

#Exemplo 2:
#função recebe: dataset mtcars,
#               info a priori "cyl"
#               colunas para coordenadas x e y "mpg" e "qsec"

disp.descritiva(mtcars, "cyl", c("mpg", "qsec"))

}
