% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq.descritiva}
\alias{freq.descritiva}
\title{Tabela de frequência}
\usage{
freq.descritiva(
  dados,
  condicional,
  vars,
  dg = FALSE,
  tabs = FALSE,
  grafs = NULL,
  grade = c(1, 1),
  titulo = FALSE
)
}
\arguments{
\item{dados}{Dataframe com ao menos duas colunas, uma com a informação a priori e outra(s) com valores para plotar histograma e boxplot}

\item{condicional}{Coluna em que está a informação conhecida}

\item{vars}{Lista com nomes das colunas em que estão os valores com os quais serão plotados o gráfico de barras e construida a tabela de frequência}

\item{dg}{Lógico, default = FALSE, TRUE para exibir resumo e gráficos incluindo dados gerais}

\item{tabs}{Lógico, default = FALSE, TRUE para exibir tabelas de frequência quando a função é executada}

\item{grafs}{vetor c(1:n) indicando os gráficos que serão exibidos com o plot customizado, o índice do gráfico corresponde a sua posição no plot geral que a função retorna}

\item{grade}{vetor c(x, y) indicando a quantidade linhas e colunas em que serão impressos os gráficos customizados definos pelo parâmetro grafs}

\item{titulo}{Lógico ou string, default = FALSE, TRUE para exibir título no plot geral no formato evento|condicional ou string no formato "título customizado"}
}
\value{
Lista com as tabelas de frequência de cada evento A dado B
}
\description{
Lê conjunto de dados e retorna uma tabela de frequência com observações, frequência absoluta e proporcional
}
\examples{

mtcars_ <- mtcars
mtcars_$am[mtcars_$am == 0] <- "automatico"
mtcars_$am[mtcars_$am == 1] <- "manual"
freq.descritiva(mtcars_, "am", c("cyl", "gear"))

}
